refactor: extract API application lifecycle
