+++ server/internal/app/routes.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"encoding/json"
